requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
pytesseract
pdfplumber
python-docx
pandas
pillow
pypdf
openpyxl
streamlit
requests